First commit
So far, the code simply starts at 0000 and runs till 9999, then goes back to 0000. It doesn't stop when the correct pin is reached, will keep going. Needs to be able to stop when it gets the correct pin, for which I need to know how the success page looks like, or program it to go while the "Sorry, ..." is still displayed.